hw1_prompting/requirements.txt
Libraries:
    google.generativeai
    pandas
    time
    re